refactor(live): extract entry pipeline into LiveEntryCoordinator (#486) (#813)
* refactor(live): extract entry pipeline into LiveEntryCoordinator (#486)

Move the live entry decision + execution path out of LiveTradingEngine
into src/engines/live/execution/entry_coordinator.py
(LiveEntryCoordinator): check_entry_conditions (signal/sizing/SL-TP
derivation) and the base-asset-locked execute_entry ->
execute_entry_locked order path (duplicate/limit guards, balance + fee
accounting, position tracking, risk re-registration, server-side
stop-loss placement, and the emergency-close fallbacks).

This is a real-money path, so the move is verbatim: the methods are
unchanged except for self.->state. against an engine backref. The
base-asset locking and ordering (#703) are preserved — execute_entry
serialises the whole entry on the symbol's lock (state._base_asset_locks)
across submit -> track -> SL placement, and the emergency-close fallback
re-acquires it re-entrantly via state._execute_exit. The two engine
methods callers mock (_execute_exit, _record_event) are invoked through
the backref so test mocks still intercept. _resolve_take_profit_pct stays
on the engine (it is also called during construction).

The engine keeps thin delegating wrappers so the loop call sites and the
public/test entry points are unchanged; it drops ~620 lines (to ~3,574).
Two close-only-mode unit tests and the AUTO_REPAY source-introspection
test are repointed to the coordinator (the layer that now owns the logic).

Full unit suite green (4047 passed); the deterministic backtest parity
fingerprint is byte-identical before and after the extraction.


* test(live): spec close-only mocks against LiveEntryEngineState

Restore the mock spec the repointed TestCloseOnlyMode tests lost when they
moved from MagicMock(spec=LiveTradingEngine) to a bare MagicMock, so an
attribute/method-name drift in the coordinator's state access fails the
test again (CODE.md). Specs against the LiveEntryEngineState protocol —
the contract the coordinator's backref implements.


---------
